pyproject.toml
[project]
name = "mar-pj3-gr4-anchor-voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # ── 기존 (STT 파이프라인) ──────────────────────────────────
    "beautifulsoup4>=4.14.3",
    "ctranslate2>=4.7.1",
    "faster-whisper>=1.2.1",
    "ffmpeg-python>=0.2.0",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "openai-whisper>=20250625",
    "pydub>=0.25.1",
    "requests>=2.33.1",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
    # ── 추가 (서비스 실행) ─────────────────────────────────────
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "streamlit>=1.33.0",
    "deepface>=0.0.93",
    "gTTS>=2.5.0",
    "google-generativeai>=0.5.0",
    "openai>=1.23.0",
    "opencv-python>=4.9.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "audio-recorder-streamlit>=0.0.10",
    "mediapipe>=0.10.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "torchaudio>=2.11.0",
    "tf-keras>=2.20.1",
    "pandas>=3.0.2",
    "jupyter>=1.1.1",
    "transformers>=5.5.0",
]





[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
clip = { git = "https://github.com/ultralytics/CLIP.git" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
